Suppress CMake external project 'update' phase
When the 'install' target is invoked during the Debian and RPM package
builds, the external project target is re-evaluated to determine if it
is stale. By default, CMake performs an 'update' command on the source
repository, which invalidates the target and causes the external project
to be re-built.

This is a problem because when the 'install' target is invoked by the
Debian/RPM build, the 'DESTDIR' variable is passed, which makes its way
into the external project as well. This interferes with the
sub-project's installation into the build directory, and results in
unexpected files getting installed into the platform package staging
directory.

This is a problem for both Debian and RPM builds, but it is only a fatal
error in RPM builds at the moment.

Using a ref which can never change and suppressing the 'update' phase of
the external project is enough to ensure that the external project is
not deemed 'stale' during the install target invocation, and avoids the
problem.

Signed-off-by: Scott K Logan <logans@cottsay.net>