Support vendor-provided pybind11 upstream
This vendor package depends on the pybind11_vendor package. That package
will conditionally build pybind11 from source if needed. A package
downstream of a vendor package usually needs to `find_package` the
upstream vendor package for the vendored payload to be discoverable by
CMake.

Since this package itself is a vendor package, the way we can accomplish
this is to find_package pybind11 on the vendored package's behalf and
specifically tell it where to look.

Signed-off-by: Scott K Logan <logans@cottsay.net>